move gov cli tests to e2e tests
