+++ tests/e2e/gov/grpc.go
@@ -1,4 +1,4 @@
-package testutil
+package gov
 
 import (
 	"fmt"
