+++ tests/e2e/gov/query.go
@@ -1,4 +1,4 @@
-package testutil
+package gov
 
 import (
 	"fmt"
